refactoring + add logging
